feat(#389): address review — accented-Latin invisibles, usize offsets, Iterable lexicon
Copilot review on #393:
- src/anomalies.rs: the invisible branch only fired with ASCII letters on both
  sides, missing a never-legitimate invisible (ZWSP, word joiner, BOM, …) between
  accented Latin letters. Split it: ZWJ/ZWNJ (legitimate joiners in Arabic/Indic/
  emoji) still require ASCII-Latin neighbours; the rest fire on any letter
  neighbour. Added a test for `é<ZWSP>à`.
- bindings/ruby/ext: return byte offsets as `usize`, not `usize as u32` — Ruby
  Integer is unbounded, so the truncation was unnecessary (matches
  find_untranslatable).
- python/disarm/_api.py: type `lexicon` as `Iterable[str]` (not `set[str]`) and
  normalise with `set(...)` before the native call, matching slugify's
  `stopwords` and the other bindings. Added list/generator test cases.

Re-verified: core 16 tests, Ruby 54 specs, Python 221 tests; clippy (both feature
sets + ext) / fmt / ruff / mypy clean.

+++ bindings/ruby/ext/disarm/src/lib.rs
@@ -321,7 +321,7 @@ fn inspect_anomalies(
 ) -> (
     bool,
     Vec<String>,
-    Vec<(String, String, u32, u32, String, String)>,
+    Vec<(String, String, usize, usize, String, String)>,
     Option<String>,
 ) {
     let lex: HashSet<String> = lexicon.into_iter().collect();
@@ -334,8 +334,8 @@ fn inspect_anomalies(
             (
                 f.kind.as_str().to_string(),
                 f.token,
-                f.start as u32,
-                f.end as u32,
+                f.start,
+                f.end,
                 f.detail,
                 reason,
             )
